requirements.txt
coverage==4.5.1
flake8==3.5.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.2.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-Marshmallow==0.9.0
marshmallow-sqlalchemy==0.15.0
flask-rest-jsonapi==0.24.0
psycopg2==2.7.4
validate_email==1.3
Flask-Cors==3.0.7
gunicorn==19.9.0
pyJWT==1.7.1
roadwatch-data-processor==0.2.7
